Updating Suisun.R
Upaded Suisun Marsh dataset to 1-14-24 version

Also, changed the parsing of the month column to use the "Datetime" column instead of the "Date". This is to avoid 9 rows in the input excel sheet (that lives on the github by itself) that has non-date data. The rest of the rows appears to be the same as the "Date" column originally used