Removed OrganismCode 81 Salvage
Removed `Chinese Mitten Crab Temp` from the Salvage database. These were collected with a different methods than regular salvage by the CVP in 1999. From "The History of the TFCT" document:

In the late-1990s, the TFCF was inundated with Chinese mitten crabs Eriocheir sinensis (Figure 16).  The first mitten crab was encountered at the TFCF on September 1996 and, in 1998, over 750,000 mitten crabs were entrained (Siegfried 1999).  In 1999, Reclamation and SLDMWA installed a crab removal traveling screen just upstream of the secondary louvers to remove mitten crabs before they could enter the holding tanks.  A high pressure spray wash system was used to remove the mitten crabs from the screen and deposit them into a hopper on the downstream side of the screen. A grain auger in the bottom of the hopper moved crabs and debris into a disposal container (White et al. 2000). By fall 2003, mitten crab numbers dropped dramatically. Although effective for debris removal during heavy debris loads (White et al. 2000, Boutwell and Sisneros 2006), the traveling screen was never used for mitten crab removal again and was used sparingly since 2003 and was dismantled in 2014 as part of the secondary channel revamp.

Also, checked change to SpeciesCode.csv, changing codes 98 and 99 to UnID. Does not change the salvage dataset at all since it relies on the code and not the name.

Updated the salvage RDA files with data downloaded 7-12-24.